Add Contact parent class
